+++ packages/decorator/src/common/decoratorManager.ts
@@ -390,11 +390,12 @@ export class DecoratorManager extends Map {
   }
 }
 
-const manager = new DecoratorManager();
+let manager = new DecoratorManager();
 if (global['MIDWAY_GLOBAL_DECORATOR_MANAGER']) {
   console.warn(
     'DecoratorManager not singleton and please check @midwayjs/decorator version by "npm ls @midwayjs/decorator"'
   );
+  manager = global['MIDWAY_GLOBAL_DECORATOR_MANAGER'];
 } else {
   global['MIDWAY_GLOBAL_DECORATOR_MANAGER'] = manager;
 }