+++ scripts/rollback/rollback-2.12.2.sh
@@ -1,0 +1,61 @@
+#!/bin/bash
+# timestamp: 1628083711264
+
+npm dist-tag add @midwayjs/egg-layer@2.12.1 latest
+npm dist-tag add @midwayjs/egg-layer@2.12.2 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.12.1 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.12.2 latest
+npm dist-tag add @midwayjs/express-layer@2.12.1 latest
+npm dist-tag add @midwayjs/express-layer@2.12.2 beta
+tnpm dist-tag add @midwayjs/express-layer@2.12.1 latest
+tnpm dist-tag add @midwayjs/express-layer@2.12.2 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.1 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.2 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.12.1 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.12.2 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.1 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.2 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.12.1 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.12.2 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.1 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.2 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.1 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.2 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.1 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.2 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.1 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.2 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.1 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.2 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.1 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.2 latest
+npm dist-tag add @midwayjs/static-layer@2.12.1 latest
+npm dist-tag add @midwayjs/static-layer@2.12.2 beta
+tnpm dist-tag add @midwayjs/static-layer@2.12.1 latest
+tnpm dist-tag add @midwayjs/static-layer@2.12.2 latest
+npm dist-tag add @midwayjs/faas@2.12.1 latest
+npm dist-tag add @midwayjs/faas@2.12.2 beta
+tnpm dist-tag add @midwayjs/faas@2.12.1 latest
+tnpm dist-tag add @midwayjs/faas@2.12.2 latest
+npm dist-tag add @midwayjs/task@2.11.6 latest
+npm dist-tag add @midwayjs/task@2.12.1 beta
+tnpm dist-tag add @midwayjs/task@2.11.6 latest
+tnpm dist-tag add @midwayjs/task@2.12.1 latest
+npm dist-tag add @midwayjs/version@2.12.1 latest
+npm dist-tag add @midwayjs/version@2.12.2 beta
+tnpm dist-tag add @midwayjs/version@2.12.1 latest
+tnpm dist-tag add @midwayjs/version@2.12.2 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.12.1 => 2.12.2
+#  - @midwayjs/express-layer: 2.12.1 => 2.12.2
+#  - @midwayjs/koa-layer: 2.12.1 => 2.12.2
+#  - @midwayjs/serverless-app: 2.12.1 => 2.12.2
+#  - @midwayjs/serverless-fc-starter: 2.12.1 => 2.12.2
+#  - @midwayjs/serverless-fc-trigger: 2.12.1 => 2.12.2
+#  - @midwayjs/serverless-scf-trigger: 2.12.1 => 2.12.2
+#  - @midwayjs/static-layer: 2.12.1 => 2.12.2
+#  - @midwayjs/faas: 2.12.1 => 2.12.2
+#  - @midwayjs/task: 2.11.6 => 2.12.2
+#  - @midwayjs/version: 2.12.1 => 2.12.2