+++ packages-serverless/serverless-app/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.3](https://github.com/midwayjs/midway/compare/v2.12.2...v2.12.3) (2021-08-09)
+
+
+### Features
+
+* vercel starter and trigger ([#1199](https://github.com/midwayjs/midway/issues/1199)) ([7d978a2](https://github.com/midwayjs/midway/commit/7d978a2bd46bf0b96c689ace2a9268e31a2e4acd))
+
+
+
+
+
+## [2.12.2](https://github.com/midwayjs/midway/compare/v2.12.1...v2.12.2) (2021-08-04)
+
+**Note:** Version bump only for package @midwayjs/serverless-app
+
+
+
+
+
 ## [2.12.1](https://github.com/midwayjs/midway/compare/v2.12.0...v2.12.1) (2021-08-01)
 
 **Note:** Version bump only for package @midwayjs/serverless-app