+++ packages-serverless/serverless-scf-trigger/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.3](https://github.com/midwayjs/midway-faas/compare/v2.12.2...v2.12.3) (2021-08-09)
+
+**Note:** Version bump only for package @midwayjs/serverless-scf-trigger
+
+
+
+
+
+## [2.12.2](https://github.com/midwayjs/midway-faas/compare/v2.12.1...v2.12.2) (2021-08-04)
+
+**Note:** Version bump only for package @midwayjs/serverless-scf-trigger
+
+
+
+
+
 ## [2.12.1](https://github.com/midwayjs/midway-faas/compare/v2.12.0...v2.12.1) (2021-08-01)
 
 **Note:** Version bump only for package @midwayjs/serverless-scf-trigger